throughput less than 1000 is seen as 0. Fixed by casting to double
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@1209 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ hazelcast-monitor/src/main/java/com/hazelcast/monitor/server/InstanceChartGenerator.java
@@ -91,7 +91,7 @@ public abstract class InstanceChartGenerator extends HttpServlet {
         TimeSeries ts = new TimeSeries("operations per second");
         for (int i = 0; i < list.size(); i++) {
             InstanceStatistics instanceStatistics = (InstanceStatistics) list.get(i);
-            ts.addOrUpdate(new Second(instanceStatistics.getCreatedDate()), instanceStatistics.getTotalOPS() / 1000);
+            ts.addOrUpdate(new Second(instanceStatistics.getCreatedDate()), (double)instanceStatistics.getTotalOPS() / 1000);
         }
         TimeSeriesCollection timeDataset = new TimeSeriesCollection();
         timeDataset.addSeries(ts);
